+++ apps/desktop/lib/i18n/strings.g.dart
@@ -4,9 +4,9 @@
 /// To regenerate, run: `dart run slang`
 ///
 /// Locales: 1
-/// Strings: 37
+/// Strings: 32
 ///
-/// Built on 2026-08-02 at 09:55 UTC
+/// Built on 2026-08-09 at 20:37 UTC
 
 // coverage:ignore-file
 // ignore_for_file: type=lint, unused_import