+++ apps/desktop/lib/ui/service/service_cubit.dart
@@ -1,7 +1,6 @@
 import 'dart:async';
 
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:remote_rift_api/remote_rift_api.dart';
 
 import '../../services/api_service_runner.dart';
 import 'service_state.dart';
@@ -22,18 +21,6 @@ class ServiceCubit extends Cubit<ServiceState> {
     await _startService();
   }
 
-  void completeStartup() async {
-    final pendingState = switch (state) {
-      PendingMultipleAddresses state => state,
-      _ => throw StateError(
-        'Tried to complete startup while not in multiple addresses state (was ${state.runtimeType})',
-      ),
-    };
-
-    emit(pendingState.produce((draft) => draft.starting = true));
-    await _startService(resolveAddress: true);
-  }
-
   void restart() async {
     final startupError = switch (state) {
       StartupError state => state,
@@ -46,14 +33,12 @@ class ServiceCubit extends Cubit<ServiceState> {
     await _startService();
   }
 
-  Future<void> _startService({bool resolveAddress = false}) async {
+  Future<void> _startService() async {
     try {
-      await runner.run(resolveAddressOnMany: resolveAddress);
+      await runner.run();
       emit(Started());
     } catch (error) {
       final newState = switch (error) {
-        MultipleAddressesFound() => PendingMultipleAddresses(starting: false),
-        AddressNotFound() => StartupError(cause: .addressNotFound),
         _ => StartupError(cause: .unknown),
       };
       emit(newState);